some functions was moved to BaseTest
